Move client into its own module
