+++ Dockerfile
@@ -6,5 +6,6 @@ COPY requirements.txt requirements.txt
 RUN pip install -r requirements.txt
 
 COPY impf.py impf.py
+COPY byimpf byimpf
 
 ENTRYPOINT ["python", "impf.py"]
